vmm: move sending memory to the other migration code
... to simplify sending memory from multiple connections in future
commits.

On-behalf-of: SAP julian.stecklina@sap.com
